+++ Vungle/VungleRouter.h
@@ -11,7 +11,6 @@
 
 extern NSString *const kVungleAppIdKey;
 extern NSString *const kVunglePlacementIdKey;
-extern NSString *const kVungleFlexViewAutoDismissSeconds;
 extern NSString *const kVungleUserId;
 extern NSString *const kVungleOrdinal;
 extern NSString *const kVungleStartMuted;